API: Add new exceptions module and move exception exposed via numeric
This means moving ComplexWarning, TooHardError, and AxisError.
