+++ numpy/core/src/umath/scalarmath.c.src
@@ -1549,7 +1549,7 @@ static PyObject *
  *
  */
 
-/* 
+/*
  * Complex numbers do not support remainder so we manually make sure that the
  * operation is not defined.  This is/was especially important for longdoubles
  * due to their tendency to recurse for some operations, see gh-18548.
@@ -1711,7 +1711,7 @@ static int
 emit_complexwarning(void)
 {
     static PyObject *cls = NULL;
-    npy_cache_import("numpy.core", "ComplexWarning", &cls);
+    npy_cache_import("numpy.exceptions", "ComplexWarning", &cls);
     if (cls == NULL) {
         return -1;
     }
